sbm - fixed warnings

+++ core/smartbody/smartbody-lib/src/sbm/me_ct_locomotion_analysis.cpp
@@ -374,7 +374,7 @@ void MeCtLocomotionAnalysis::analyze_limb_anim(MeCtLocomotionLimbAnim* anim, SkM
 			}
 		}
 
-		displacement /= anim->get_support_joint_num();
+		displacement /= (float)anim->get_support_joint_num();
 	}
 
 	// calculated the velocity if the base joint is fixed
